Conflicts:
	Acr.XamForms.Autofac/Acr.XamForms.Autofac.csproj
	Acr.XamForms.Autofac/packages.config
	Acr.XamForms.Mobile/Acr.XamForms.Mobile.csproj
	Acr.XamForms.Mobile/packages.config
	Acr.XamForms.UserDialogs.iOS/UserDialogService.cs
	Acr.XamForms.sln
	Acr.XamForms/Acr.XamForms.csproj
	Samples/Samples.iOS/Samples.iOS.csproj
	Samples/Samples/Samples.csproj
	Samples/Samples/ViewModels/UserDialogViewModel.cs
	Samples/Samples/Views/HomeView.cs
	Samples/Samples/Views/ListView.cs
	Samples/Samples/Views/MailView.xaml.cs
	Samples/Samples/Views/SettingsView.xaml
	Samples/Samples/Views/SignatureListView.xaml
	Samples/Samples/packages.config
	nuspec/Acr.XamForms.Mobile.nuspec
	nuspec/Acr.XamForms.UserDialogs.nuspec

+++ Shared Utils/Ios.cs
@@ -1,7 +1,7 @@
 #if __IOS__
 using System;
 using System.Linq;
-using MonoTouch.UIKit;
+using UIKit;
 
 
 namespace Acr.XamForms {
@@ -14,7 +14,7 @@ namespace Acr.XamForms {
                 .Windows
                 .Reverse()
                 .FirstOrDefault(x => 
-                    x.WindowLevel == UIWindowLevel.Normal && 
+(float)                    x.WindowLevel == (float)UIWindowLevel.Normal && 
                     !x.Hidden
                 );
